requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
bcrypt==4.3.0
build==1.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
config==0.5.1
cryptography==45.0.5
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.7.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meraki==2.0.3
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
netmiko==4.6.0
ntc_templates==7.9.0
numpy==2.3.1
openai==1.97.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
paramiko==3.5.1
pluggy==1.6.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyfiglet==1.0.3
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
pyserial==3.5
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scp==0.15.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
textfsm==1.1.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
utils==1.0.2
wheel==0.45.1
yarl==1.20.1
